feat: extract downloadAndDecryptAttachments to own util file
